+++ src/pudl/metadata/fields.py
@@ -1087,6 +1087,13 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "boolean",
         "description": "Indicate whether the generator can deliver power to the transmission grid.",
     },
+    "delivered_energy_mmbtu": {  # created for eiaaeo
+        "type": "number",
+        "description": (
+            "Energy delivered to the point of use, as a proxy for energy consumption. Includes fuels and electric power; excludes losses from generation and transmission."
+        ),
+        "unit": "MMBtu",
+    },
     "delivered_mwh": {
         "type": "number",
         "description": "Gross megawatt-hours delivered in power exchanges and used as the basis for settlement.",
@@ -3940,9 +3947,27 @@ FIELD_METADATA: dict[str, dict[str, Any]] = {
         "type": "string",
         "description": "Identifier indicating original FERC Form 1 source record. format: {table_name}_{report_year}_{report_prd}_{respondent_id}_{spplmnt_num}_{row_number}. Unique within FERC Form 1 DB tables which are not row-mapped.",
     },
+    "region_name_eiaaeo": {
+        "type": "string",
+        "description": (
+            "EIA AEO region for energy consumption. Includes US Census Divisions plus United States."
+        ),
+    },
     "region_name_us_census": {
         "type": "string",
         "description": "Human-readable name of a US Census region.",
+    },
+    "region_type_eiaaeo": {
+        "type": "string",
+        "description": (
+            "Region type for EIA AEO energy consumption, indicating whether region_name_eiaaeo is a US Census Division or country (United States)"
+        ),
+        "constraints": {
+            "enum": [
+                "us_census_division",
+                "country",
+            ],
+        },
     },
     "has_regulatory_limits": {
         "type": "boolean",